[mlir][emitc] Deferred emission as expressions
The translator currently implements deferred emission for certain ops.
Like expressions, these ops are emitted as part of their users but
unlike expressions, this is mandatory. Besides complicating the code
with a second inlining mechanism, deferred emission's inlining is
limited as it's not recursive.

This patch extends EmitC's expressions to deferred emission ops by (a)
marking them as CExpressions, (b) extending expression interface to
mark ops as always-inline and (c) support inlining of always-inline
CExpressions even when not packed of an `emitc.expression` op,
retaining current behavior.